dev-python/django-tastypie: fix broken "doc" dependencies
The "doc" conditional dependencies were, strangely, different from the
runtime dependency and pinned specifically on django 1.9 which had to be
masked for security reasons.

Because of that, we had to mask the "doc" USE flag on the package.

In this commit, we remove these strange dependencies and use the same as
the runtime deps. Documentation builds fine.

Committing as straight-to-stable because documentation simply couldn't
be built before that.

Package-Manager: Portage-2.3.43, Repoman-2.3.10