x11-libs/wxGTK: new slot - 3.1-gtk3
This 3.1 slot is required by software like http://cubicsdr.com/

A 3.1.1 subslot is added to allow for ":SLOT=" dependencies in packages
that uses this development version of the library in which the ABI can
change between patch or minor versions.

A CubicSDR package depending on "x11-libs/wxGTK:3.1-gtk3=[opengl]"
will be rebuilt whenever the wxGTK subslot changes.

Closes: https://bugs.gentoo.org/577030
Closes: https://github.com/gentoo/gentoo/pull/9359
Authored-by: Ștefan Talpalaru <stefantalpalaru@yahoo.com>
Signed-off-by: Ștefan Talpalaru <stefantalpalaru@yahoo.com>