net-im/telegram-desktop-bin: exclude from session management
The application seems to register with the X11 session manager but fails
to follow the protocol for remembering its state. The result is that a
session manager will restart it, and during that process the command
line argument for disabling the internal updater gets lost.
So on a restarted session the updater will be enabled again.

Do not let the application register at the session manager in the first
place, by unsetting the env variable used to find it.

Closes: https://bugs.gentoo.org/662330
Signed-off-by: Henning Schild <henning@hennsch.de>