media-libs/gst-plugins-base: add media-libs/graphene using revision
GL elements automagically use graphene. But we want them always anyways,
if GL is enabled, so add a revision that just forces the dependency.
This is a separate revision, as it involves dropping keywords and it
would cascade into dozens of packages for other 1.14.2 version bumps,
thus keep a 1.14.2-r0 without it forced to avoid that for now.
If newly packaged graphene is installed separately, then 1.14.2-r0
will have an automagic dependency on it, but this here is the work
to fix that.

Package-Manager: Portage-2.3.43, Repoman-2.3.10