net-libs/gupnp-1.0.3: py3 support and python-single-r1.eclass
Python is used for install gupnp-binding-tool /usr/bin utility,
there's no point installing it for all python versions supported,
even if more than one is supported, as there's no point for the
/usr/bin wrapper. Use python-single-r1 properly instead.
The package not working with python3 seems to be an old comment,
and other distributions seem to use it fine, so support up to
py3.6 as well as the single-target choice. I don't have 3.7 quite
yet to test that.

Package-Manager: Portage-2.3.43, Repoman-2.3.10