dev-python/django-compressor: maintainer needed
The package hasn't been updated in a while and has no revdev. Nobody in
the Python team is interested in maintaining it. It depends on an
unsupported version of django and unless maintained will go away soon.

In this commit, I make the <django-1.9 dependency explicit and drop
support for python3_6 (not supported by django 1.8)

Package-Manager: Portage-2.3.44, Repoman-2.3.10