dev-python/fabric: Bump to 2.2.1
The jump for 1.x to 2.x upstream is a major change so I've added a new
slot "2" for it along with a new "fab2" USE flag to decide whether we
install in "side-by-side" mode.

I didn't go for an eselect script because this side-by-side thing is a
transition measure that shouldn't last very long. I also went with the
"fab2" USE flag because it was what was the closest to upstream's
transition strategy[1]. Also:

* EAPI 7
* Python 3.6
* Temporarily drop ~arm64 because of new "invoke" dep
* Remove broken bash completions

In addition to not being able to make bash completions work under 1.x,
the completion situation changed with fab 2.x with invoke being able to
generate them automatically but I haven't been able to make it work for
fabric yet.

[1]: http://www.fabfile.org/installing.html

Closes: https://bugs.gentoo.org/657392
Package-Manager: Portage-2.3.43, Repoman-2.3.10