media-sound/ncmpc: version bump to 0.30
Ebuild for ncmpc-0.30

Haven't test lirc - only building with it installed.

Also, there appears to be a bug with going to song from search window
where is goes to wrong song or says the song is not in the playlist.
Need to be confirmed and fixed before commiting this ebuild.

media-sound/ncmpc: version bump to 0.30

Haven't test lirc - only building with it installed.

Also, there appears to be a bug with going to song from search window
where is goes to wrong song or says the song is not in the playlist.
Need to be confirmed and fixed before commiting this ebuild.

media-sound/ncmpc: bugfix

Fix for enqueue from search list going to the wrong song.

media-sound/ncmpc: ebuild cleanup

media-sound/ncmpc: remove trailing newline

media-sound/ncmpc: replace trailing whitespaces with a tab

media-sound/ncmpc: call einstalldocs unconditionally

media-sound/ncmpc: fixes for review issues