media-video/cheese: bump to 3.28.0, remove unnecessary vala dep
Remove unnecessary vala dep; tarball comes with pregenerated C and
does not even use valac as we build it (as C code is present and
regeneration with system vala version is not triggered).
Also workaround unnecessary glib[utils] requirement: upstream
configure checks for gtester-report and disables tests as a whole
when not found, when actually the default "make check" target
doesn't use it; thus just make the configure check be happy
without gtester-report present and remove the USE dep.

Package-Manager: Portage-2.3.43, Repoman-2.3.10