profiles/embedded/make.defaults: drop some PROFILE_ONLY_VARIABLES
ARCH, ELIBC, and KERNEL variables need to be set or overridden when using
the embedded profile for special crossdev projects (i.e. mingw32/64).
Crossdev tools often set the overrides already in make.conf, but they were
previously ignored when the embedded profile is used.  Also of note, none
of those variables are set in the embedded profile itself.  This commit allows
for those variables to be defined in make.conf, so that the minimalist
embedded profile can be used directly rather than needing to create a custom
profile based on it instead.

Bug: https://bugs.gentoo.org/487310