dev-python/pillow: add qt5 USE flag
Pillow has a Qt compatibility layer that is enabled depending on whether
PyQt is installed or not. Letting this happen can lead to unexpected
problems, so we add support for the qt5 flag to explicitly enable Qt
support. When the flag is disabled, we forcibly disable Qt support
through patching.

Bug: https://bugs.gentoo.org/662236
Package-Manager: Portage-2.3.44, Repoman-2.3.10