dev-java/oracle-jdk-bin-10.0.2 version bump
Copied from oracle-jdk-bin-9.0.4-r2.ebuild with one change.

A Orace-JVM-feature called "usage tracker" was introduced with some version of 1.8 with default disabled. But with 10, it sees that it become enabled by default. This will result in the creation of $USER/.java_usagetracker/, which will be prevented by the sandbox.

The best way here IMHO is to use a special command line property, see: https://docs.oracle.com/javacomponents/usage-tracker/overview/ . Please notice the change at line 198 for that.

But for the later usage, the user may (IMHO should) block this by a property file located at .../conf/management (former .../lib/management)

Bug: https://bugs.gentoo.org/651040
Signed-off-by: Guido Jäkel G.Jaekel@DNB.DE