mail-client/thunderbird: bump to v60.0
This is the initial version of the v60 series of thunderbird.

Once started, thunderbird profiles cannot be reverted to work with older
versions of thunderbird!  This has never been guaranteed, but
it is absolutely known to break with this update.

This is the first version of thunderbird using the GTK3 toolkit, as such
there may be some artifacts or issues with some dialogs or other UI
elements.

As with newer firefox, this version of thunderbird requires rust to build,
and so is limited in the number of arches it can be built with.

Package-Manager: Portage-2.3.40, Repoman-2.3.9