net-analyzer/nagios-icinga-openvpn: new package to monitor OpenVPN.
This package is intended to be a full-featured replacement for the
existing net-analyzer/nagios-check_openvpn-simple plugin. The "simple"
version lacked several important features, the most important of which
was the ability to use a TLS-auth key in the default UDP configuration.

The new plugin is largely a drop-in replacement, but the executable
name is different, so your configuration will need to be updated.

Bug: https://bugs.gentoo.org/663774
Package-Manager: Portage-2.3.44, Repoman-2.3.9