dev-util/glib-utils: drop python2 from compat
python2 handles non-utf8 files badly for glib-mkenums, while
it seems to be just a warning with python3. So drop python2
support, at least for now.
Before it was kept due to bad glib[$PYTHON_USEDEP] deps, when
mkenums was part of dev-libs/glib, but that's not a concern
with this new package.

Bug: https://bugs.gentoo.org/647892
Package-Manager: Portage-2.3.47, Repoman-2.3.10