net-analyzer/nload: don't compress the man-page
As pointed out by @mgorny [1], packages should not compress their man pages on their own.
Instead, Portage needs to handle the compression.
Currently, `ecompressdir` is smart enough to recompress the man-pages as needed.
However, this is not a fully PMS-compliant behavior.

So, patch the Makefile.am not to compress the man-page.
A corresponding pull request is submitted upstream [2].

[1] https://github.com/gentoo/gentoo/pull/9543#issuecomment-415662844
[2] https://github.com/rolandriegel/nload/pull/4

Closes: https://github.com/gentoo/gentoo/pull/9543