media-plugins/calf: work around buggy configure check
The upstream project's configure.ac hard-codes a check to see if calf is
already installed to /usr/lib or /usr/local/lib, and if so, refuses to
configure (causing econf to fail) if `calf` is already installed on the
system.

Simply adding --without-obsolete-check to the configure invocation
disables this (pointless) check.

Closes: https://bugs.gentoo.org/661036
Closes: https://github.com/gentoo/gentoo/pull/9701
Signed-off-by: Niklas Haas <git@haasn.xyz>
Signed-off-by: Miroslav Šulc <fordfrog@gentoo.org>