mail-client/thunderbird-bin: bump to v60.0
This is a major release bump.  Although various security bugs are fixed
in this release, it may not be stable enough to replace 52.9 and so
runtime testing is recommended.  Note specifically that upgrading to
this release will change the thunderbird profile in ways that are not
backwards-compatible with 52.x; back up your ~/.thunderbird if you
wish to be able to downgrade.

Package-Manager: Portage-2.3.40, Repoman-2.3.9