use.desc: Introduce 'gtk-doc' global USE flag
gtk-doc API docs are not shipped pregenerated with meson anymore, thus
developers might want to enable these kind of docs globally, independent
of everything else that USE=doc covers. Therefore introduce a separate
USE flag for this directly as a global USE flag, as many meson ports, to
be committed soon, will make use of it instead of USE=doc.
This is meant for cases where gtk-doc is otherwise not installed.
Typically autotools based build systems pregenerate gtk-docs into disted
tarball and such a USE flag is not necessary there, as the configure
argument makes it only regenerate it.

Link: https://archives.gentoo.org/gentoo-dev/message/f399ebe56c25ab895eb36a236747d76e