net-wireless/wpa_supplicant: make dependency on CRDA optional
The crda tool can be used to send information about a wireless
regulatory domain to the kernel to apply regulations depending on the
country the user is currently in. As it is possible to compile the
database directly into the kernel by providing "regulatory.db" from
net-wireless/wireless-regdb via CONFIG_EXTRA_FIRMWARE, a userspace crda
executable is not always required.

Add a "+crda" useflag that toggles the dependency on net-wireless/crda.
As crda currently requires Python 2 due to the dependency on m2crypto,
deactivating the flag can result in quite some savings.