app-emulation/crossover-bin: Bump to 17.5.1
This also includes a fix for the install process breaking when a recent
version of konqueror is installed. Crossover relies on it for KDE
version detection while Konqueror in its infinite wisdom decided to open
things with write permissions, which are sandbox violations. Right now,
the fix is only in 17.5.1. I need to talk to upstream and think about
this some more before I decide the proper course of action.

Package-Manager: Portage-2.3.40, Repoman-2.3.9