dev-games/ogre: Pre 2.1 release snapshot 2018-09-01 added
As the future of Ogre changes a lot of the old API, it is important to have
access to the new system as early as possible. The current development state
is quite usable already, although not all components have been ported, yet.

Providing an ebuild for the new 2.1 series to interested users and developers,
everybody gets the chance to adapt their project to the future of Ogre.

This is done using snapshot ebuilds, because not all commits in the Ogre tree
provide buildable revisions. Further it is easier to test newly ported
components when snapshot ebuilds are used.

Bug: https://bugs.gentoo.org/647266
Closes: https://bugs.gentoo.org/647266
Signed-off-by: Sven Eden <sven.eden@prydeworx.com>
Package-Manager: Portage-2.3.48, Repoman-2.3.10