net-misc/networkmanager: bump to 1.12.2, support iwd
iwd support is always built with USE=wifi, as net-wireless/iwd
itself is not necessary for build due to pure dbus usage.
USE=iwd controls whether USE=wifi ensures wpa_supplicant or iwd
is present at runtime, and adds a default config snippet to use
iwd as wifi.backend instead of wpa_supplicant.
As a result, iwd and wpa_supplicant can be used interchangably
with configuration overrides in /etc, despite what NM iwd USE
setting was on the install, as long as USE=wifi was enabled; but
a choice based default is ensured to be working in lack of overrides.

Also remove unused versionator and toolchain-funcs inherits and
add now needed glib-utils build dep directly for glib migration
to that.

Closes: https://bugs.gentoo.org/661278
Package-Manager: Portage-2.3.48, Repoman-2.3.10