dev-python/pillow: re-support pypy
pypy support had been dropped in pillow 4.3 ebuild and is being
re-introduced now. Tests pass.

This required a little bit of library path fiddling in setup.py, the
root cause of it being that `sysconfig.get_config_var('prefix')` under
pypy doesn't return the same thing as with CPython. This fiddling was
required if we wanted to keep the --disable-platform-guessing flag.

I tried not to break the fix that was done in bug 661830.

I've tried pypy3, but not all tests passed. I didn't investigate further
(I'm just trying to return to pre-4.3 levels so that we can stabilize)

Bug: https://bugs.gentoo.org/661830
Closes: https://bugs.gentoo.org/608620
Package-Manager: Portage-2.3.49, Repoman-2.3.10