dev-util/boost-build: add patch to unbreak compiler options on sparc
Boost tries to autodetect the processor architecture and overrides any flags
set with CXXFLAGS with the -mcpu value it thinks it is appropiate. Sadly the
most recent architecture it knows of is ultrasparc3. For every newer cpu type
it falls back to it's default, which is v7 (which would be the compiler default
anyway). This avoids any advanced cpu instructions, e.g. those that support
atomic operations on thing larger than a byte. Remove the whole outdated cruft
and just use whatever the user passed in.

Closes: https://bugs.gentoo.org/646234
Closes: https://github.com/gentoo/gentoo/pull/9802
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>
Package-Manager: Portage-2.3.40, Repoman-2.3.9