python-utils-r1.eclass: New python_fix_shebang approach
The previous approach would erroneously match foopython. The new
approach requires the match to start the string or be preceeded by a
slash, the only two cases we actually want. It does this with slightly
less code and allows the replacement of whole path strings that would
be problematic when passed to sed. This will be needed when
cross-compiling is addressed.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>