sys-devel/binutils: expand SONAME with USE=multitarget, bug #666100
In #663690 creative use of LD_LIBRARY_PATH exposed a problem
of Gentoo's split across binutils and binutils-libs:
we install libbfd.so with the same SONAME but different ABIs.
This caused 'ld' to SIGSEGV or fail asserts.

The change here is to give unique SONAME to each of:
native binutils, cross binutils, binutils-libs packages.

On top of that we encode USE=multitarget flags into SONAME as well.
binutils-libs will also include USE=64-bit-bfd.

This way dynamic linker will be able to detect incompatible changes
like:
- rebuild binutils with new USE flags
- install binutils-libs (silently overrides binutils' libbfd.so)
and report it explicitly.

Bug: https://bugs.gentoo.org/666100
Package-Manager: Portage-2.3.49, Repoman-2.3.10