cuda.eclass: update to EAPI7, remove 0-4
cuda.eclass is only used from EAPI5,6 currently so remove support for
older EAPIs.

Updating to EAPI7 means removing versionator. It was only used to sort
the cuda versions. Instead first try the current GCC version, if that
fails try best_version, if that still fails just pick the last in
the list that works. They are already sorted and hopefully stay that
way so version_sort is not required.

Signed-off-by: Jason Zaman <perfinion@gentoo.org>