mail-filter/spamassassin: new version 3.4.2.
This new version incorporates a ton of fixes that have piled on top of
v3.4.1 over the past few years. As a result, we are able to drop all
of our PATCHES in the new ebuild. The new version also addresses four
security vulnerabilities:

  * CVE-2016-1238
  * CVE-2017-15705
  * CVE-2018-11780
  * CVE-2018-11781

Sadly, there is a new failure in the test suite (upstream bug 7622) so
I've added RESTRICT=test for now.

Stabilization of this version should already be fast-tracked for the
security issues mentioned above, but just in case, the release notes
have the following to say:

  If you do not update to 3.4.2, you will be stuck at the last ruleset
  with SHA-1 signatures in the near future.

Bug: https://bugs.gentoo.org/666348
Bug: https://bz.apache.org/SpamAssassin/show_bug.cgi?id=7622
Package-Manager: Portage-2.3.40, Repoman-2.3.9