mail-filter/spamassassin: disable a network test and drop RESTRICT=test.
I originally added RESTRICT=test to spamassassin-3.4.2.ebuild because
a test was failing, I didn't know why, and we needed to get the new
version out to address multiple CVEs. It turns out that the one
failing test is not respecting the fact that network access should be
disabled in the test suite (upstream bug 7622).

In the new revision, we kill that one test in src_prepare(). The rest
of the tests should pass, so RESTRICT=test has been removed.

Bug: https://bugs.gentoo.org/666348
Bug: https://bz.apache.org/SpamAssassin/show_bug.cgi?id=7622
Package-Manager: Portage-2.3.40, Repoman-2.3.9