Revert "app-emulation/libvirt: update live ebuild"
This reverts commit 2ad130bab3b38de59ff996dad8f3a1c68407e731.

So after some painful bugs and v.4.6.0 release where we tried to
make libvirt work with jansson we came to realize that it is not
going to fly. QEMU relies on sending 64 bit integers and jansson
supports only 53 bits long integers. Therefore we reverted back
to using yajl. See d99a89592d65ab8 in libvirt repo (and friends).

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Signed-off-by: Matthias Maier <tamiko@gentoo.org>