dev-lang/luajit: Fix 2.0.5 cross-compilation build
luajit passes CFLAGS/LDFLAGS to both host and target builds.
This breaks when the host cpu does not support the target flags
e.g. using march=bdver4 in CFLAGS when cross-compiling for AMD
on an Intel host.

Instead, pass the flags in {HOST|TARGET}-{CFLAGS|LDFLAGS} which is
supported by luajit [1].

[1] https://github.com/LuaJIT/LuaJIT/blob/master/doc/install.html#L607

Closes: https://github.com/gentoo/gentoo/pull/9970
Signed-off-by: Manoj Gupta <manojgupta@google.com>
Signed-off-by: William Hubbs <williamh@gentoo.org>