media-libs/libplacebo: Add new ebuild
New project for video rendering that VLC 4 plans to use more heavily,
and which also be used by mpv in the foreseeable future.

Incidentally, this is an optional minor dependency of VLC 3, so we could
also add libplacebo support to that ebuild (which improves the color
management of the opengl vout).

Signed-off-by: Niklas Haas <git@haasn.xyz>