www-client/epiphany: bump to 3.30.0 and p.masked 3.30.1
The MSE enabling in 3.30.1 to fix having Youtube playback for over 360p
is crashing hard for me, unable to even look at 240p. So this is package
masked until some more investigations. It's supposed to be experimental,
but just outright crashing sounds bad and I got tired of sitting in
coredumpctl service runs for minutes.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.49, Repoman-2.3.11