app-text/mupdf: bump to 1.14.0
Many things changed upstream in this release, which prompts many changes
in the ebuild.

1. Drop the "static" USE flag. It was already broken in 1.13 because
options weren't passed to the static `emake` call like is was in the
main `emake` invocation. In addition to that, building a static binary
requires "media-gfx/graphite2[+static-libs]" which doesn't exist.

2. Drop "lcms" USE flag. See bug #667788.

3. Always link to dev-lang/mujs. mupdf can't build without being linked
to mujs now. The best we can do it to disable JS with the "FZ_ENABLE_JS"
flag, which is what the "javascript" USE flag does.

4. Make "javascript" USE flag enabled by default, which is in line with
upstream's behavior.

5. Make "vanilla" USE flag enabled by default. The patch it enables
doesn't seem very useful and I couldn't track it's *raison d'être* in
git/cvs history: it's been there since the first commit.

6. Simplify ebuild and remove obsolete seds. The "sed" invocations I've
removed have been verified and either point to absent files or replace
nothing.

7. Don't remove third party libs. The new "USE_SYSTEM_LIBS" flag being
used is a reliable way to ensure that the correct libraries are used.

8. Fix "curl" USE flag. See bug #664646

Closes: https://bugs.gentoo.org/664646
Closes: https://bugs.gentoo.org/667788

Signed-off-by: Virgil Dupras <vdupras@gentoo.org>
Package-Manager: Portage-2.3.50, Repoman-2.3.11