media-gfx/sane-backends: Fixes from Fedora and Debian...
They fix an important bug with Samsung devices not being properly recognized
and noises in some devices.

We also start disabling locking as all the other distributions and stop
trying to randomly generate extra documentation files that are not really
needed and add a lot of hidden dependencies tend to break in several ways.

Closes: https://bugs.gentoo.org/636202
Closes: https://bugs.gentoo.org/653300
Closes: https://bugs.gentoo.org/668232
Closes: https://bugs.gentoo.org/668350
Signed-off-by: Pacho Ramos <pacho@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11