net-libs/webkit-gtk: add gst-plugins-opus dependency for MSE
MSE needs opus, though techically this is optional. Pull it in
with USE=gstreamer always for now, to trigger the opusparse
related blocker in epiphany for now, while not actually depending
on new enough gst-plugins-opus here, as we are ahead in keywords
for webkit due to security.
Later this dep should be removed again (once gst-plugins-opus
with opusparse is stable) or the dep raised to 1.14.4-r1.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.49, Repoman-2.3.11