dev-java/openjdk: New package to provide Java 11 from source
I'm not sure if this will build on hardened. There used to be an
upstream mechanism to handle PaX markings but this disappeared in an
earlier version. It might build with USE=-jbootstrap though. Please
let us know.

Thanks to Alex Dubov for the initial ebuild attempt.

Closes: https://github.com/gentoo/gentoo/pull/9835
Signed-off-by: James Le Cuirot <chewi@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11