x11-libs/wxGTK: update patchset and fix USE=sdl dep
configure checks for sdl2 first, and unconditionally picks that
if found, instead of libsdl. So as we are in 2018, just force
libsdl2 for USE=sdl to guarantee that what is depended on is
also picked.

Update patchset to early November state for a number of wxGTK
bug fixes found on upstream WX_3_0_BRANCH.

Closes: https://bugs.gentoo.org/669414
Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.49, Repoman-2.3.11