dev-util/netsurf-buildsystem: new package
This replaces the netsurf eclass which not necessary and too complex for
its own need. Being an eclass makes modifying it too heavy a process.

Most of netsurf.eclass' purpose was to wrap netsurf's build system,
which is now what we do in this package. In addition to installing it as
a normal package, we add a helper shell script that can then be loaded
in consumer ebuilds.

Signed-off-by: Virgil Dupras <vdupras@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11