www-client/netsurf: backtrack on the dep bundling idea
The dep bundling idea introduced in 3.8-r1 wasn't a very good idea
after all, as we can see in #669714. A better idea, which is what is
being done in the commits preceding this one, is to drop netsurf eclass
in all dependencies by introducing dev-util/netsurf-buildsystem as a
regular package.

Bug: https://bugs.gentoo.org/669714
Signed-off-by: Virgil Dupras <vdupras@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11