toolchain.eclass: drop info/dir index, bug #672408
As explained by Zac in #672408 prepinfo() was used to drop 'info/dir'
index to allow portage regenerate it.

gcc package does not always install 'info/dir' files (USE-dependent).
This causes nondeterminism for 'info/dir' to be owned or be an orphan
file.

This change drops 'info/dir' unconditionally to avoid owned/orphan
collisions and always make it an orphan file.

Reported-by: Michal Jakubowski
Bug: https://bugs.gentoo.org/672408
Bug: https://bugs.gentoo.org/587316
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>