app-admin/sudo: Add secure-path use
It makes sense to mask PATH variable when sudo is executed for secure
reasons with known secure content. Problem is that this is not common on
other distributions and in some cases it makes sense to allow user's
PATH to be used because it complicates common use of sudo such as:
sudo !!
This does not change default previous behavior. It just adds an use flag
to enabling secure-path that is in default enabled.

Bug: https://bugs.gentoo.org/672522

Signed-off-by: Karel Kočí <cynerd@email.cz>