app-text/gtranslator: lock gnome-dictionary subslot
Newer gnome-dictionary will lose subslot and not ship a public
libgdict anymore, thus will be unsuitable for this gtranslator
gnome-dictionary support. The pre-existing := in VDB should
catch the rebuild need and then find it won't be fine with the
newer gnome-dictionary (which will lose the subslot = subslot 0
instead of 10). Presumably newer gtranslator loses the libgdict
optional linking.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.52, Repoman-2.3.11