x11-libs/gtksourceview: bump to 3.24.9
Remove unecessary itstool dep - configure checks for it, but
it's non-fatal and nothing uses the resulting ITSTOOL build
variable.
Default enable USE=vala, because various things are starting
to require it in a meson build world - already the case with
gnome-calculator.
Stop passing --disable-deprecations, as that actually adds
G_DISABLE_DEPRECATED and so on, which we actually do NOT want
to happen - but we ended up not having them anyways, as we
sed the configure logic to not add them. So this instructs
it to do the right thing, instead of relying on the eclass
logic to force us doing the right thing.

Bug: https://bugs.gentoo.org/665578
Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.52, Repoman-2.3.11