dev-libs/folks: fix problem building against tracker-2
Introspection files must be regenerated since they contain
hard-references to the installed tracker revision which causes problems
on tracker soversion changes like the 2.0 release. The easiest way to
do that is to re-generate sources from vala directly. This problem
should vanish if the project is ever ported to meson.

Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Gilles Dartiguelongue <eva@gentoo.org>