gnome-base/nautilus-3.24.2.1: preparation for tracker-2
* Lock tracker to tracker-1, as configure doesn't support
tracker-2.
* Remove nautilus-tracker-tags PDEPEND, as it is rather
unmaintained and going away with tracker-2. Meanwhile, old
tracker has a USE=nautilus that pulls it in anyways, and
another way to enable this in nautilus shouldn't have
existed in the first place. This change is mainly meant to
allow us to hopefully last rite nautilus-tracker-tags
sooner, as all this combined warranted a revbump anyways,
thus getting the deps into user VDBs.

Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.52, Repoman-2.3.11