net-libs/tox: strip vulnerable versions
This commit removes all versions of tox prior to
version 0.2.8 because those versions were vulnerable
to remote DDoS attacks. Also, this commit removes a
deprecated use flag in metadata.xml. In addition,
this commit adds the requirement that libsodium
not be built with the minimal use flag enabled in
tox-9999.

Closes: https://bugs.gentoo.org/628530
Bug: https://bugs.gentoo.org/668264
Signed-off-by: Josiah Mullins <JoMull01@protonmail.com>
Signed-off-by: Josiah Mullins <jomull01@protonmail.com>
Closes: https://github.com/gentoo/gentoo/pull/10585
Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>