gnome-base/librsvg: bump to 2.40.20
Give up hope of having backported to 2.40.19 test suite machinery
work with its numerous issues (too old freetype generated reference
images, issues for 32-bit, and so on and on), restrict tests and
roll with it. This probably has some security relevant last fixes
for the non-rust version and we should have it, even if tests fail.
Ideally tests will work with the rust requiring versions (on the
architectures that can support, anyways).

Signed-off-by: Mart Raudsepp <leio@gentoo.org>
Package-Manager: Portage-2.3.52, Repoman-2.3.11