games-board/teg: add support for teg configure options.
The client, debug, ggz, robot, and server USE flag options are added.
The htmltidy package is obsolete, so the respective dependency is
replaced with a tidy-html5 dependency. In addition, the metadata.xml
file is updated with a long description of the project, descriptions for
the new flags, and a switch to the Github upstream remote id. The
project release is now pulled from the project Github page, where the
project moved from Sourceforge.

Closes: https://bugs.gentoo.org/671448
Signed-off-by: William Breathitt Gray <vilhelm.gray@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/10668
Signed-off-by: Patrice Clement <monsieurp@gentoo.org>