profiles/prefix/darwin/macos/arch/ppc: mask gcc and binutils-apple versions
gcc-8.2.0 compiles and runs, but seems to cause issues with Perl's
Configure -- since it's still out there kind of experimentally, mask it
for now.

binutils-apple-6.3 looked like a viable path to use a more recent
compiler, but causes bus errors during gcc compilations, so avoid it for
now, to be revisited once we have a more stable environment.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>