games-emulation/dosbox: Make X11 dependency optional
The x11-libs/libX11 dependency was missing although this was
transitively covered by libsdl[X]. It's not really required though and
DOSBox detects it somewhat automagically. If libX11 is present but
libsdl was not built with X support then the build actually fails. For
these reasons, we have to forcibly disable X11 using an autoconf cache
variable.

Package-Manager: Portage-2.3.53, Repoman-2.3.12
Signed-off-by: James Le Cuirot <chewi@gentoo.org>