gnome-extra/libgda: workaround build failures without vala
Upstream configure has a bad check for the libgee not existing failure,
by checking a variable that's not used anymore, so the check will
always fail with no libgee found - no matter if vala extensions
are actually requested or not. Due to older version not building
against new openssl, workaround it quickly and temporarily by
always requiring libgee during build, pending a proper fix that
fixes the check in configure.

Bug: https://bugs.gentoo.org/674066
Package-Manager: Portage-2.3.52, Repoman-2.3.12
Signed-off-by: Mart Raudsepp <leio@gentoo.org>