dev-perl/Devel-Cover: Move dep for B::Debug to dev-perl/B-Debug
Devel-Cover is the only package in tree that indicates it requires
B::Debug. More over, no packages we have in gentoo turn up on CPAN
when grepped for consumption of B::Debug

Therefor, we're just going to skip the shenigans with virtuals and make
all consumers ( who we remeber, are all in ~arch ) simply pull
dev-perl/B-Debug

This won't cause any problems with older virtual versions existing,
as no package exists requiring a specific version of B-Debug, so the
multi-way conflict virtuals typically do with perl-core/ are not
necessary here.

So, Devel-Cover is safe to force installing a CPAN version of B-Debug
on any perl.

Bug: https://bugs.gentoo.org/661202
Package-Manager: Portage-2.3.51, Repoman-2.3.11
Signed-off-by: Kent Fredric <kentnl@gentoo.org>